photo-studio-manager/requirements.txt
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
reportlab==4.0.4
openpyxl==3.1.2
Pillow==10.0.1
python-dateutil==2.8.2
mysql-connector-python==8.1.0
